Lift Negative onto a UnaryOperator base class
Negative held its operand, its location, its printing and its equality
itself, all of which are facts about prefix operators rather than about
negation. BinaryOperator already draws that line for infix operators;
this is its counterpart, and what is left in Negative is what actually
differs between one prefix operator and another: the token it is spelled
with, how it evaluates, and the type it gives the result.

Printing follows the token, the way it now does for binary operators:
Precedence::unary() is the same shape as Precedence::binary(), except
that its operand sits at the unary level itself rather than one tighter,
because parseUnary() reads its operand by calling itself. That also makes
Precedence::parenthesize() private — Negative was its last outside
caller.

The base class carries no @internal, and can't: Negative published every
member it now inherits, and Psalm resolves internality from the declaring
class. What that costs is written down where it binds, along with the
issue that undoes it at the next major.
